+++ macros/Cargo.toml
@@ -3,7 +3,7 @@ name = "wasm-bindgen-utils-macros"
 version = "0.0.6"
 edition = "2021"
 license = "CAL-1.0"
-description = "Provides helper macros for wasm-bindgen-utils"
+description = "Provides helper proc macros for wasm-bindgen-utils"
 homepage = "https://github.com/rainlanguage/rain.wasm"
 
 [lib]
