+++ macros/src/wasm_export/mod.rs
@@ -1,14 +1,93 @@
+use quote::quote;
+use proc_macro::TokenStream;
+use syn::{Error, ImplItem, ItemImpl};
+
 // Module for WASM export functionality
-mod function_gen;
-mod transformers;
-mod type_utils;
+mod attrs;
+mod fn_tools;
 
 // Re-export the public API
-pub use function_gen::create_new_function_call;
-pub use transformers::{collect_function_arguments, handle_attrs};
-pub use type_utils::try_extract_result_inner_type;
+pub use fn_tools::*;
+pub use attrs::handle_attrs;
 
 // Constants used throughout the module
-pub const WASM_EXPORT_ATTR: &str = "wasm_export";
 pub const SKIP_PARAM: &str = "skip";
+pub const WASM_EXPORT_ATTR: &str = "wasm_export";
 pub const UNCHECKED_RETURN_TYPE_PARAM: &str = "unchecked_return_type";
+
+/// Starts macro parsing and expansion process
+pub fn expand(_attr: TokenStream, item: TokenStream) -> Result<TokenStream, Error> {
+    // Parse the input as an impl block
+    let mut input = syn::parse::<ItemImpl>(item)?;
+
+    // Create vector to store exported items
+    let mut export_items = Vec::new();
+
+    for item in input.items.iter_mut() {
+        if let ImplItem::Fn(method) = item {
+            // Process for export if applicable
+            if let syn::Visibility::Public(_) = method.vis {
+                let (forwarding_attrs, inner_ret_type, should_skip) = handle_attrs(method)?;
+                if should_skip {
+                    continue;
+                }
+                if let Some(inner_ret_type) = inner_ret_type {
+                    let fn_name = &method.sig.ident;
+                    let is_async = method.sig.asyncness.is_some();
+                    let (has_self_receiver, args) = collect_function_arguments(&method.sig.inputs);
+
+                    // Create exported version
+                    let export_fn_name = syn::Ident::new(
+                        &format!("{}__{}", fn_name, WASM_EXPORT_ATTR),
+                        fn_name.span(),
+                    );
+
+                    let mut export_method = method.clone();
+                    export_method.sig.ident = export_fn_name;
+                    export_method
+                        .attrs
+                        .push(syn::parse_quote!(#[allow(non_snake_case)]));
+                    export_method.attrs.extend(forwarding_attrs);
+
+                    let new_return_type = syn::parse_quote!(-> WasmEncodedResult<#inner_ret_type>);
+                    export_method.sig.output = new_return_type;
+
+                    let call_expr = create_new_function_call(fn_name, has_self_receiver, &args);
+
+                    if is_async {
+                        export_method.block = syn::parse_quote!({
+                            #call_expr.await.into()
+                        });
+                    } else {
+                        export_method.block = syn::parse_quote!({
+                            #call_expr.into()
+                        });
+                    }
+
+                    export_items.push(ImplItem::Fn(export_method));
+                } else {
+                    return Err(Error::new_spanned(
+                        &method.sig.output,
+                        "expected Result<T, E> return type",
+                    ));
+                }
+            }
+        }
+    }
+
+    // Create two impl blocks
+    let original_impl = input.clone();
+
+    let mut export_impl = input;
+    export_impl.items = export_items;
+
+    // Generate the output with wasm_bindgen only on the export impl
+    let output = quote! {
+        #original_impl
+
+        #[wasm_bindgen]
+        #export_impl
+    };
+
+    Ok(output.into())
+}
